requirements.txt
# ============================================================================
# Portia Bot - Requirements
# ============================================================================
# FILE VERSION: v1.0.0
# Repository: https://github.com/the-alphabet-cartel/portia
# Community: The Alphabet Cartel - https://discord.gg/alphabetcartel
# ============================================================================

fluxer-py
